[EC-428] Moved CreateProviderCommand to Commercial.Core project
